+++ MelonLoader/Libs/SafeEnumParse.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using System.Security;
+using System.Text;
+
+namespace MelonLoader
+{
+    internal static class SafeEnumParse
+    {
+
+    }
+}